+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Fixed object diff rest changes notes
+- Fixed context hook paths duplication in case of re-renders
 
 ## 0.4.0 (September 23, 2021)